Changes to polygon payment support

+++ core/payment-driver/base/migrations/2021-10-13-033400_payment_driver_improvements/up.sql
@@ -12,7 +12,7 @@ CREATE TABLE `transaction_tmp`(
     current_gas_price DOUBLE NULL,
     limit_gas_price DOUBLE NULL,
     final_gas_price DOUBLE NULL,
-    final_gas_used DOUBLE NULL,
+    final_gas_used INTEGER NULL,
     gas_limit INTEGER NULL,
     time_created DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     time_last_action DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
